AI_Model/chat_bot/requirements.txt

streamlit
python-dotenv
google-generativeai
pydantic